+++ lib/toolbar/widgets/toolbar.dart
@@ -6,7 +6,7 @@ import '../../documents/models/attribute.model.dart';
 import '../../shared/models/editor-dialog-theme.model.dart';
 import '../../shared/models/editor-icon-theme.model.dart';
 import '../../shared/widgets/arrow-scrollable-button-list.dart';
-import '../../shared/widgets/quill-icon-button.dart';
+import '../../shared/widgets/icon-button.dart';
 import '../models/editor-custom-icon.dart';
 import '../models/media-picker.type.dart';
 import 'buttons/camera-button.dart';
@@ -24,7 +24,7 @@ import 'buttons/toggle-style-button.dart';
 import 'buttons/video-button.dart';
 
 export '../../embeds/services/image-video.utils.dart';
-export '../../shared/widgets/quill-icon-button.dart';
+export '../../shared/widgets/icon-button.dart';
 export 'buttons/clear-format-button.dart';
 export 'buttons/color-button.dart';
 export 'buttons/history-button.dart';
@@ -38,7 +38,7 @@ export 'buttons/toggle-style-button.dart';
 export 'buttons/video-button.dart';
 
 // The default size of the icon of a button.
-const defaultIconSize = 18.0;
+const defaultIconSize = 21.0;
 
 // The factor of how much larger the button is in relation to the icon.
 const iconButtonFactor = 1.77;
@@ -92,7 +92,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
   factory EditorToolbar.basic({
     required EditorController controller,
     double toolbarIconSize = defaultIconSize,
-    double toolbarSectionSpacing = 4,
+    double toolbarSectionSpacing = 2.5,
     WrapAlignment toolbarIconAlignment = WrapAlignment.center,
     bool showDividers = true,
     bool showFontSize = true,
@@ -188,9 +188,20 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
           '48': 48
         };
 
+    Widget _divider() => Container(
+          margin: EdgeInsets.only(
+            top: multiRowsDisplay ? 6.5 : 0,
+            left: 7,
+            right: 7,
+          ),
+          height: toolbarIconSize + 5,
+          width: 1.3,
+          color: Colors.grey.shade400,
+        );
+
     return EditorToolbar(
       key: key,
-      toolbarHeight: toolbarIconSize * 2,
+      toolbarHeight: toolbarIconSize * 3,
       buttonsSpacing: toolbarSectionSpacing,
       toolbarIconAlignment: toolbarIconAlignment,
       multiRowsDisplay: multiRowsDisplay,
@@ -202,6 +213,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.undo_outlined,
             iconSize: toolbarIconSize,
             controller: controller,
+            buttonsSpacing: toolbarSectionSpacing,
             undo: true,
             iconTheme: iconTheme,
           ),
@@ -210,6 +222,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.redo_outlined,
             iconSize: toolbarIconSize,
             controller: controller,
+            buttonsSpacing: toolbarSectionSpacing,
             undo: false,
             iconTheme: iconTheme,
           ),
@@ -247,6 +260,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
           ToggleStyleButton(
             attribute: AttributeM.bold,
             icon: Icons.format_bold,
+            buttonsSpacing: toolbarSectionSpacing,
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
@@ -255,6 +269,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
           ToggleStyleButton(
             attribute: AttributeM.italic,
             icon: Icons.format_italic,
+            buttonsSpacing: toolbarSectionSpacing,
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
@@ -263,6 +278,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
           ToggleStyleButton(
             attribute: AttributeM.small,
             icon: Icons.format_size,
+            buttonsSpacing: toolbarSectionSpacing,
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
@@ -273,6 +289,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.format_underline,
             iconSize: toolbarIconSize,
             controller: controller,
+            buttonsSpacing: toolbarSectionSpacing,
             iconTheme: iconTheme,
           ),
         if (showStrikeThrough)
@@ -281,6 +298,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.format_strikethrough,
             iconSize: toolbarIconSize,
             controller: controller,
+            buttonsSpacing: toolbarSectionSpacing,
             iconTheme: iconTheme,
           ),
         if (showInlineCode)
@@ -289,6 +307,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.code,
             iconSize: toolbarIconSize,
             controller: controller,
+            buttonsSpacing: toolbarSectionSpacing,
             iconTheme: iconTheme,
           ),
         if (showColorButton)
@@ -296,6 +315,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.color_lens,
             iconSize: toolbarIconSize,
             controller: controller,
+            buttonsSpacing: toolbarSectionSpacing,
             background: false,
             iconTheme: iconTheme,
           ),
@@ -303,6 +323,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
           ColorButton(
             icon: Icons.format_color_fill,
             iconSize: toolbarIconSize,
+            buttonsSpacing: toolbarSectionSpacing,
             controller: controller,
             background: true,
             iconTheme: iconTheme,
@@ -310,6 +331,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
         if (showClearFormat)
           ClearFormatButton(
             icon: Icons.format_clear,
+            buttonsSpacing: toolbarSectionSpacing,
             iconSize: toolbarIconSize,
             controller: controller,
             iconTheme: iconTheme,
@@ -318,6 +340,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
           ImageButton(
             icon: Icons.image,
             iconSize: toolbarIconSize,
+            buttonsSpacing: toolbarSectionSpacing,
             controller: controller,
             onImagePickCallback: onImagePickCallback,
             filePickImpl: filePickImpl,
@@ -330,6 +353,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
           VideoButton(
             icon: Icons.movie_creation,
             iconSize: toolbarIconSize,
+            buttonsSpacing: toolbarSectionSpacing,
             controller: controller,
             onVideoPickCallback: onVideoPickCallback,
             filePickImpl: filePickImpl,
@@ -349,6 +373,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             filePickImpl: filePickImpl,
             webImagePickImpl: webImagePickImpl,
             webVideoPickImpl: webVideoPickImpl,
+            buttonsSpacing: toolbarSectionSpacing,
             iconTheme: iconTheme,
           ),
         if (showDividers &&
@@ -358,16 +383,13 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
                 isButtonGroupShown[3] ||
                 isButtonGroupShown[4] ||
                 isButtonGroupShown[5]))
-          VerticalDivider(
-            indent: 12,
-            endIndent: 12,
-            color: Colors.grey.shade400,
-          ),
+          _divider(),
         if (showAlignmentButtons)
           SelectAlignmentButton(
             controller: controller,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
+            buttonsSpacing: toolbarSectionSpacing,
             showLeftAlignment: showLeftAlignment,
             showCenterAlignment: showCenterAlignment,
             showRightAlignment: showRightAlignment,
@@ -376,6 +398,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
         if (showDirection)
           ToggleStyleButton(
             attribute: AttributeM.rtl,
+            buttonsSpacing: toolbarSectionSpacing,
             controller: controller,
             icon: Icons.format_textdirection_r_to_l,
             iconSize: toolbarIconSize,
@@ -387,14 +410,11 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
                 isButtonGroupShown[3] ||
                 isButtonGroupShown[4] ||
                 isButtonGroupShown[5]))
-          VerticalDivider(
-            indent: 12,
-            endIndent: 12,
-            color: Colors.grey.shade400,
-          ),
+          _divider(),
         if (showHeaderStyle)
           SelectHeaderStyleButton(
             controller: controller,
+            buttonsSpacing: toolbarSectionSpacing,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
           ),
@@ -404,13 +424,10 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             (isButtonGroupShown[3] ||
                 isButtonGroupShown[4] ||
                 isButtonGroupShown[5]))
-          VerticalDivider(
-            indent: 12,
-            endIndent: 12,
-            color: Colors.grey.shade400,
-          ),
+          _divider(),
         if (showListNumbers)
           ToggleStyleButton(
+            buttonsSpacing: toolbarSectionSpacing,
             attribute: AttributeM.ol,
             controller: controller,
             icon: Icons.format_list_numbered,
@@ -422,6 +439,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             attribute: AttributeM.ul,
             controller: controller,
             icon: Icons.format_list_bulleted,
+            buttonsSpacing: toolbarSectionSpacing,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
           ),
@@ -430,6 +448,7 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             attribute: AttributeM.unchecked,
             controller: controller,
             icon: Icons.check_box,
+            buttonsSpacing: toolbarSectionSpacing,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
           ),
@@ -438,21 +457,19 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             attribute: AttributeM.codeBlock,
             controller: controller,
             icon: Icons.code,
+            buttonsSpacing: toolbarSectionSpacing,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
           ),
         if (showDividers &&
             isButtonGroupShown[3] &&
             (isButtonGroupShown[4] || isButtonGroupShown[5]))
-          VerticalDivider(
-            indent: 12,
-            endIndent: 12,
-            color: Colors.grey.shade400,
-          ),
+          _divider(),
         if (showQuote)
           ToggleStyleButton(
             attribute: AttributeM.blockQuote,
             controller: controller,
+            buttonsSpacing: toolbarSectionSpacing,
             icon: Icons.format_quote,
             iconSize: toolbarIconSize,
             iconTheme: iconTheme,
@@ -463,44 +480,43 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
             iconSize: toolbarIconSize,
             controller: controller,
             isIncrease: true,
+            buttonsSpacing: toolbarSectionSpacing,
             iconTheme: iconTheme,
           ),
         if (showIndent)
           IndentButton(
             icon: Icons.format_indent_decrease,
+            buttonsSpacing: toolbarSectionSpacing,
             iconSize: toolbarIconSize,
             controller: controller,
             isIncrease: false,
             iconTheme: iconTheme,
           ),
         if (showDividers && isButtonGroupShown[4] && isButtonGroupShown[5])
-          VerticalDivider(
-            indent: 12,
-            endIndent: 12,
-            color: Colors.grey.shade400,
-          ),
+          _divider(),
         if (showLink)
           LinkStyleButton(
             controller: controller,
             iconSize: toolbarIconSize,
+            buttonsSpacing: toolbarSectionSpacing,
             iconTheme: iconTheme,
             dialogTheme: dialogTheme,
           ),
         if (customIcons.isNotEmpty)
-          if (showDividers)
-            VerticalDivider(
-              indent: 12,
-              endIndent: 12,
-              color: Colors.grey.shade400,
-            ),
+          if (showDividers) _divider(),
         for (var customIcon in customIcons)
           IconBtn(
-              highlightElevation: 0,
-              hoverElevation: 0,
-              size: toolbarIconSize * iconButtonFactor,
-              icon: Icon(customIcon.icon, size: toolbarIconSize),
-              borderRadius: iconTheme?.borderRadius ?? 2,
-              onPressed: customIcon.onTap),
+            highlightElevation: 0,
+            buttonsSpacing: toolbarSectionSpacing,
+            hoverElevation: 0,
+            size: toolbarIconSize * iconButtonFactor,
+            icon: Icon(
+              customIcon.icon,
+              size: toolbarIconSize,
+            ),
+            borderRadius: iconTheme?.borderRadius ?? 2,
+            onPressed: customIcon.onTap,
+          ),
       ],
     );
   }
@@ -510,11 +526,13 @@ class EditorToolbar extends StatelessWidget implements PreferredSizeWidget {
     return I18n(
       initialLocale: locale,
       child: multiRowsDisplay
-          ? Wrap(
-              alignment: toolbarIconAlignment,
-              runSpacing: 4,
-              spacing: buttonsSpacing,
-              children: children,
+          ? Container(
+              padding: EdgeInsets.symmetric(vertical: 10),
+              child: Wrap(
+                alignment: toolbarIconAlignment,
+                runSpacing: 4,
+                children: children,
+              ),
             )
           : Container(
               constraints: BoxConstraints.tightFor(